requirements.txt
websockets
azure.cognitiveservices.speech
pydub